Split annotator to multiple files
