+++ salk_toolkit/commands.py
@@ -49,7 +49,7 @@ def streamlit_fn_factory(relpath: str, curpath: str) -> Callable[[], None]:
 # Run explorer app
 run_explorer = streamlit_fn_factory("./tools/explorer.py", os.path.dirname(__file__))
 # Run annotator app
-run_annotator = streamlit_fn_factory("./tools/annotator.py", os.path.dirname(__file__))
+run_annotator = streamlit_fn_factory("./tools/annotator/annotator.py", os.path.dirname(__file__))
 
 
 def translate_pot(template: str, dest: str, t_func: Callable[[str], str], sources: list[str] | None = None) -> None:
