+++ salk_toolkit/tools/annotator/__init__.py
@@ -1,0 +1,1 @@
+"""Streamlit annotation editor (split into modules)."""
